Move BACKGROUND_TEXT_MODE from design-system to elements package
