+++ packages/design-system/src/utils/constants.js
@@ -33,12 +33,6 @@ export const FOCUSABLE_SELECTORS = [
   '[tabindex]:not([tabindex="-1"])',
 ];
 
-export const BACKGROUND_TEXT_MODE = {
-  NONE: 'NONE',
-  FILL: 'FILL',
-  HIGHLIGHT: 'HIGHLIGHT',
-};
-
 export const OverlayType = {
   NONE: 'none',
   SOLID: 'solid',
