+++ packages/element-library/src/text/display.js
@@ -25,12 +25,14 @@ import { useUnits } from '@googleforcreators/units';
 import { useTransformHandler } from '@googleforcreators/transform';
 import { getHTMLFormatters, getHTMLInfo } from '@googleforcreators/rich-text';
 import { stripHTML } from '@googleforcreators/dom';
-import { BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
 import {
   getResponsiveBorder,
   shouldDisplayBorder,
 } from '@googleforcreators/masks';
-import { StoryPropTypes } from '@googleforcreators/elements';
+import {
+  StoryPropTypes,
+  BACKGROUND_TEXT_MODE,
+} from '@googleforcreators/elements';
 
 /**
  * Internal dependencies
