+++ packages/elements/src/constants.js
@@ -72,3 +72,9 @@ export const TEXT_ELEMENT_DEFAULT_FONT = {
     lGap: 0,
   },
 };
+
+export const BACKGROUND_TEXT_MODE = {
+  NONE: 'NONE',
+  FILL: 'FILL',
+  HIGHLIGHT: 'HIGHLIGHT',
+};
