+++ packages/story-editor/src/components/library/panes/text/useInsertPreset.js
@@ -24,7 +24,7 @@ import {
 } from '@googleforcreators/react';
 import { dataFontEm, PAGE_HEIGHT } from '@googleforcreators/units';
 import { getHTMLFormatters } from '@googleforcreators/rich-text';
-import { BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
+import { BACKGROUND_TEXT_MODE } from '@googleforcreators/elements';
 import { calculateTextHeight } from '@googleforcreators/element-library';
 
 /**
