+++ packages/story-editor/src/components/panels/design/textStyle/backgroundColor.js
@@ -21,10 +21,8 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { useCallback, useRef } from '@googleforcreators/react';
 import { __ } from '@googleforcreators/i18n';
-import {
-  PillGroup,
-  BACKGROUND_TEXT_MODE,
-} from '@googleforcreators/design-system';
+import { PillGroup } from '@googleforcreators/design-system';
+import { BACKGROUND_TEXT_MODE } from '@googleforcreators/elements';
 
 /**
  * Internal dependencies
