+++ packages/story-editor/src/components/styleManager/styleItem.js
@@ -21,7 +21,8 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { __ } from '@googleforcreators/i18n';
 import { generatePatternStyles } from '@googleforcreators/patterns';
-import { Icons, BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
+import { Icons } from '@googleforcreators/design-system';
+import { BACKGROUND_TEXT_MODE } from '@googleforcreators/elements';
 import { useState } from '@googleforcreators/react';
 import { stripHTML } from '@googleforcreators/dom';
