+++ packages/story-editor/src/utils/presetUtils.js
@@ -25,7 +25,7 @@ import {
 } from '@googleforcreators/patterns';
 import { getHTMLInfo } from '@googleforcreators/rich-text';
 import { generateFontFamily } from '@googleforcreators/element-library';
-import { BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
+import { BACKGROUND_TEXT_MODE } from '@googleforcreators/elements';
 
 /**
  * Internal dependencies
