+++ packages/story-editor/src/utils/test/presetUtils.js
@@ -17,8 +17,10 @@
 /**
  * External dependencies
  */
-import { BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
-import { TEXT_ELEMENT_DEFAULT_FONT } from '@googleforcreators/elements';
+import {
+  TEXT_ELEMENT_DEFAULT_FONT,
+  BACKGROUND_TEXT_MODE,
+} from '@googleforcreators/elements';
 
 /**
  * Internal dependencies
