+++ packages/story-editor/src/utils/useApplyTextAutoStyle.js
@@ -18,7 +18,7 @@
  * External dependencies
  */
 import { getHTMLFormatters } from '@googleforcreators/rich-text';
-import { BACKGROUND_TEXT_MODE } from '@googleforcreators/design-system';
+import { BACKGROUND_TEXT_MODE } from '@googleforcreators/elements';
 
 /**
  * Internal dependencies
